Ajout des listes de coordonnees

+++ ocr/coord.h
@@ -2,11 +2,11 @@
 # define COORD_H_
 # include "matrix.h"
 struct coord{
-		  int maxup;
-		  int maxdown;
-		  int maxleft;
-		  int maxright;
-		  int numnote;
+		  size_t maxup;
+		  size_t maxdown;
+		  size_t maxleft;
+		  size_t maxright;
+		  size_t numnote;
 };
 struct s_matrix* drawrect(struct s_matrix *mat,struct coord * coord);
